fix(bedrock): complete 1.26.32 hook wiring and correct enum underlying types
- Update spawnResources/_sendMessage symbol names in the win/linux configs for
  the new 1.26.32 signatures (actorContext param, const-ref filtered message)
- Thread actorContext through the block-drop call sites (campfire, leaves)
- Correct pre-existing enum underlying types against the r26_u3 headers
  (AbilitiesIndex, CreativeItemCategory, Flip, BossEventUpdateType,
  DBHelpers::Category, ArmSize::Type)

+++ src/bedrock/world/actor/player/abilities.h
@@ -19,7 +19,7 @@
 
 #include "bedrock/bedrock.h"
 
-enum class AbilitiesIndex : std::int16_t {
+enum class AbilitiesIndex : std::int8_t {
     Invalid = -1,
     Build = 0,
     Mine = 1,
